validattion and db insert/update with auth tok working
