+++ db/user_man.js
@@ -1,6 +1,11 @@
 const Datastore = require('nedb');
+const val = require("./validation.js")
 
 // Create collection
+upd_opts = {
+	upsert: true,
+}
+
 options = {
 	filename : '../db/user_collection',
 	timestampData : true
@@ -8,10 +13,31 @@ options = {
 
 db = new Datastore(options);
 
-// Create index settings
-index_opts = {
-	fieldName: 'tracks.id',
-	unique: 'true',
+// db.ensureIndex(index_opts, (err)=>{});
+db.persistence.setAutocompactionInterval(30*1000);
+
+db.loadDatabase ((err) => {console.log(err); return;});
+
+function update_token (user, expire_time)
+{
+    let unix_time = Math.floor(new Date() / 1000);
+    user.expire_time = unix_time + expire_time;
+
+    // console.log(user);
+
+    if(val.validate(user))
+    {
+        db.update({ _id: user._id }, user, { upsert: true }, (err, numReplaced) => {
+                if(err){
+                    console.log(err);
+                    return false;
+                }
+                console.log
+                return true;
+            } );
+    }
 }
 
-pl_db.ensureIndex(index_opts, (err)=>{});
+module.exports = {
+	update_token: update_token
+}
