+++ package.json
@@ -10,6 +10,7 @@
     "debug": "^4.3.1",
     "express": "~4.16.0",
     "got": "^11.8.1",
+    "jsonschema": "^1.4.0",
     "lodash": "^4.17.20",
     "nedb": "^1.8.0",
     "os": "^0.1.1",
